requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements/test.in
#
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
click==8.1.7
    # via code-annotations
code-annotations==1.8.0
    # via -r requirements/test.in
coverage[toml]==7.6.4
    # via pytest-cov
django==4.2.16
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
exceptiongroup==1.2.2
    # via pytest
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via code-annotations
markupsafe==3.0.2
    # via jinja2
packaging==24.2
    # via pytest
pbr==6.1.0
    # via stevedore
pluggy==1.5.0
    # via pytest
pytest==8.3.3
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==6.0.0
    # via -r requirements/test.in
pytest-django==4.9.0
    # via -r requirements/test.in
python-slugify==8.0.4
    # via code-annotations
pyyaml==6.0.2
    # via code-annotations
sqlparse==0.5.1
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.3.0
    # via code-annotations
text-unidecode==1.3
    # via python-slugify
tomli==2.0.2
    # via
    #   coverage
    #   pytest
typing-extensions==4.12.2
    # via
    #   -r requirements/base.txt
    #   asgiref
